backend/requirements.txt
fastapi
uvicorn
pandas
pandasai
sqlalchemy
asyncpg
python-dotenv
google-generativeai
pydantic
python-multipart
